Ajustando o banco de dados

+++ SistemaLojadeRoupas/backend/src/main/resources/db/migration/V1_2__Create_Table_Product.sql
@@ -1,7 +1,7 @@
 CREATE TABLE public.tb_product (
 	id SERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    price DOUBLE,
+    price INTEGER,
     size TEXT,
     quantity INTEGER,
     color TEXT,
